pyproject.toml
[project]
name = "fitness-tracker-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.1.0",
    "python-telegram-bot>=22.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "mypy>=1.16.1",
    "python-lsp-server[all]>=1.12.2",
]
